Shift components around. Moved header component out of BaseLayout component and added to App.jsx. Add flight stage label with colours to header. Added router for '/' for main dashboard and '/grid' for grid dashboard view (placeholder)
