+++ frontend/src/components/charts/TemperatureChart.jsx
@@ -8,12 +8,12 @@ import {
   Label,
   ResponsiveContainer
 } from 'recharts';
-import formatDateTick from '../services/formatDateTick';
-import Loader from '../services/Loader';
+import formatDateTick from '../../services/formatDateTick';
+import Loader from '../../services/Loader';
 import ChartContainer from './ChartContainer';
 
-export default function HumidityChart({ humidityData, handleChartClick }) {
-  const data = humidityData.slice(-50);
+export default function TemperatureChart({ tempData, handleChartClick }) {
+  const data = tempData.slice(-50);
 
   // Custom tooltip formatter (optional, for better formatting)
   const CustomTooltip = ({ active, payload, label }) => {
@@ -36,7 +36,7 @@ export default function HumidityChart({ humidityData, handleChartClick }) {
           }}
         >
           <p className='label'>{`Time: ${time}`}</p>
-          <p className='data'>{`Humidity: ${payload[0].value.toFixed(2)}%`}</p>
+          <p className='data'>{`Temperature: ${payload[0].value.toFixed(2)}°C`}</p>
         </div>
       );
     }
@@ -45,16 +45,16 @@ export default function HumidityChart({ humidityData, handleChartClick }) {
   };
 
   return (
-    <ChartContainer onZoom={handleChartClick('humidity')} title='Humidity/Time'>
-      {humidityData.length == 0 ? (
+    <ChartContainer title='Temperature/Time' onZoom={handleChartClick('temperature')}>
+      {tempData.length == 0 ? (
         <Loader />
       ) : (
-        <ResponsiveContainer width='100%' height='100%'>
+        <ResponsiveContainer width='100%' height='90%'>
           <LineChart
             width={730}
             height={250}
             data={data}
-            margin={{ top: 5, right: 30, left: 20, bottom: 30 }}
+            margin={{ top: 5, right: 30, left: 20, bottom: 10 }}
           >
             <CartesianGrid strokeDasharray='3 3' stroke='#5d5e5e' />
             <XAxis
@@ -65,16 +65,15 @@ export default function HumidityChart({ humidityData, handleChartClick }) {
               stroke='white'
             />
             <YAxis
-              dataKey='humidity'
+              dataKey='temperature'
               tick={{ fill: 'white', dx: -10, fontSize: 12 }}
-              angle={0}
               stroke='white'
             >
-              <Label value={'Humidity (%)'} angle={-90} fill='white' dx={-45} />
+              <Label value={'Temperature (°C)'} angle={-90} fill='white' dx={-45} />
             </YAxis>
-            <Tooltip content={CustomTooltip} />
+            <Tooltip content={<CustomTooltip />} />
             {/* <Legend /> */}
-            <Line type='monotone' dataKey='humidity' stroke='#fff' dot={false} />
+            <Line type='monotone' dataKey='temperature' stroke='#fff' dot={false} />
           </LineChart>
         </ResponsiveContainer>
       )}
