+++ frontend/src/components/charts/TermperatureChartFull.jsx
@@ -5,57 +5,40 @@ import {
   YAxis,
   CartesianGrid,
   Tooltip,
-  Legend,
   Label,
-  ResponsiveContainer,
-  ReferenceArea
+  ResponsiveContainer
 } from 'recharts';
-import formatDateTick from '../services/formatDateTick';
+import formatDateTick from '../../services/formatDateTick';
 import ChartContainer from './ChartContainer';
 
 const calculateMovingAverage = (data, windowSize) => {
   const movingAverageData = [];
 
   for (let i = 0; i < data.length; i += 5) {
-    let acxSum = 0;
-    let acySum = 0;
-    let aczSum = 0;
+    let sum = 0;
     let count = 0;
 
     for (let j = Math.max(0, i - windowSize + 1); j <= i; j++) {
-      acxSum += data[j].acx;
-      acySum += data[j].acy;
-      aczSum += data[j].acz;
+      sum += data[j].temperature;
       count++;
     }
 
-    const acxAverage = acxSum / count;
-    const acyAverage = acySum / count;
-    const aczAverage = aczSum / count;
-
-    movingAverageData.push({
-      dateTime: data[i].dateTime,
-      acx: acxAverage,
-      acy: acyAverage,
-      acz: aczAverage
-    });
+    const average = sum / count;
+    movingAverageData.push({ dateTime: data[i].dateTime, temperature: average });
   }
 
   return movingAverageData;
 };
 
-export default function AccelerationChartFull({ accelerationData }) {
-    
-
-  // console.log(spectData);
-  if (!accelerationData) {
+export default function TemperatureChartFull({ temperatureData }) {
+  if (!temperatureData) {
     // Loading placeholder so things don't break
     return <div>Loading</div>;
   }
 
-  const data = accelerationData;
+  const data = temperatureData;
 
-  const movingAverageData = calculateMovingAverage(accelerationData, 50);
+  const movingAverageData = calculateMovingAverage(temperatureData, 10);
 
   // Custom tooltip formatter (for better formatting)
   const CustomTooltip = ({ active, payload, label }) => {
@@ -78,9 +61,7 @@ export default function AccelerationChartFull({ accelerationData }) {
           }}
         >
           <p className='label'>{`Time: ${time}`}</p>
-          <p className='data'>{`X: ${payload[0].value.toFixed(2)}`}</p>
-          <p className='data'>{`Y: ${payload[1].value.toFixed(2)}`}</p>
-          <p className='data'>{`Z: ${payload[2].value.toFixed(2)}`}</p>
+          <p className='data'>{`Temperature: ${payload[0].value.toFixed(2)}°C`}</p>
         </div>
       );
     }
@@ -89,7 +70,7 @@ export default function AccelerationChartFull({ accelerationData }) {
   };
 
   return (
-    <ChartContainer title='Acceleration/Time'>
+    <ChartContainer title='Temperature/Time'>
       <ResponsiveContainer width='100%' height='95%'>
         <LineChart
           width={730}
@@ -105,18 +86,17 @@ export default function AccelerationChartFull({ accelerationData }) {
             interval={Math.ceil(data.length / 100)}
             stroke='white'
           />
-          <YAxis tick={{ fill: 'white', dx: -10, fontSize: 16 }} angle={0} stroke='white'>
-            <Label value={'Acceleration'} angle={-90} fill='white' dx={-30} />
+          <YAxis
+            dataKey='temperature'
+            tick={{ fill: 'white', dx: -5, fontSize: 16 }}
+            angle={0}
+            stroke='white'
+          >
+            <Label value={'Temperature (°C)'} angle={-90} fill='white' dx={-40} />
           </YAxis>
           <Tooltip content={CustomTooltip} />
-          <Legend
-            verticalAlign='bottom'
-            align='center'
-            wrapperStyle={{ paddingTop: '40px' }} // Add padding to separate the legend from the chart
-          />
-          <Line type='monotone' dataKey='acx' name='X-axis' stroke='#FFFF00' dot={false} />
-          <Line type='monotone' dataKey='acy' name='Y-axis' stroke='#fc0f03' dot={false} />
-          <Line type='monotone' dataKey='acz' name='Z-axis' stroke='#c603fc' dot={false} />
+          {/* <Legend /> */}
+          <Line type='monotone' dataKey='temperature' stroke='#fff' dot={false} />
         </LineChart>
       </ResponsiveContainer>
     </ChartContainer>
